==> models/00_GBDT_check.ipynb <==
import sys,os
import pandas as pd
import numpy as np
from time import time

#GBDTModel
from sklearn.ensemble import GradientBoostingClassifier
from sklearn.model_selection import cross_val_score,RepeatedStratifiedKFold
from sklearn.experimental import enable_hist_gradient_boosting
from sklearn.ensemble import HistGradientBoostingClassifier
from xgboost import XGBClassifier
from lightgbm import LGBMClassifier
from catboost import CatBoostClassifier
%precision 3


SEED = 42


df_train = pd.read_csv("./df/EDA_09_GBDT_train.csv")
df_test = pd.read_csv("./df/EDA_09_GBDT_test.csv")
Target = 'Perished'
data_cleaner = [df_train, df_test]
df_train.columns


for dataset in data_cleaner:
    dataset = dataset.drop(columns=["Name", "Sex", "Age", "IsAlone", "Ticket", "SibSp", "Parch",
                           "FareBin_Code", "Cabin", "Embarked", "Title", "FareBin", "AgeBin"], inplace=True)


USE_Columns =df_train.columns
X = df_train.iloc[:, 2:].values
y = df_train.iloc[:, 1].values
X_test = df_test.iloc[:, 1:]


accuracy = {}
speed = {}

model = GradientBoostingClassifier()
name = "GradientBoosting"
start = time()
cv = RepeatedStratifiedKFold(n_splits=5,n_repeats=3,random_state=SEED)
score = cross_val_score(model,X,y,scoring="accuracy",cv=cv,n_jobs=-1)
speed[name] = np.round(time()-start,3)
accuracy[name] = np.mean(score).round(3)
print(
    f"Mean Accuracy: {accuracy[name]}\nStd: {np.std(score):.3f}\nRun time: {speed[name]}s"
)

model = HistGradientBoostingClassifier()
name = "HistGradientBoostingClassifier"
start = time()
cv = RepeatedStratifiedKFold(n_splits=5, n_repeats=3, random_state=SEED)
score = cross_val_score(model, X, y, scoring="accuracy", cv=cv, n_jobs=-1)
speed[name] = np.round(time()-start, 3)
accuracy[name] = np.mean(score).round(3)
print("Model:{name}\n")
print(
    f"Mean Accuracy: {accuracy[name]}\nStd: {np.std(score):.3f}\nRun time: {speed[name]}s"
)


model = XGBClassifier()
name = "XGBClassifier"
start = time()
cv = RepeatedStratifiedKFold(n_splits=5, n_repeats=3, random_state=SEED)
score = cross_val_score(model, X, y, scoring="accuracy", cv=cv, n_jobs=-1)
speed[name] = np.round(time()-start, 3)
accuracy[name] = np.mean(score).round(3)
print("Model:{}\n".format(name))
print(
    f"Mean Accuracy: {accuracy[name]}\nStd: {np.std(score):.3f}\nRun time: {speed[name]}s"
)


model = LGBMClassifier()
name = "LGBMClassifier"
start = time()
cv = RepeatedStratifiedKFold(n_splits=5, n_repeats=3, random_state=SEED)
score = cross_val_score(model, X, y, scoring="accuracy", cv=cv, n_jobs=-1)
speed[name] = np.round(time()-start, 3)
accuracy[name] = np.mean(score).round(3)
print("Model:{}\n".format(name))
print(
    f"Mean Accuracy: {accuracy[name]}\nStd: {np.std(score):.3f}\nRun time: {speed[name]}s"
)


model = CatBoostClassifier()
name = "CatBoostClassifier"
start = time()
cv = RepeatedStratifiedKFold(n_splits=5, n_repeats=3, random_state=SEED)
score = cross_val_score(model, X, y, scoring="accuracy", cv=cv, n_jobs=-1)
speed[name] = np.round(time()-start, 3)
accuracy[name] = np.mean(score).round(3)
print("Model:{}\n".format(name))
print(
    f"Mean Accuracy: {accuracy[name]}\nStd: {np.std(score):.3f}\nRun time: {speed[name]}s"
)


accuracy
