==> models/00_RondomForest_check.ipynb <==
import sys
import os
import pandas as pd
import numpy as np
from time import time

#GBDTModel
from sklearn.ensemble import RandomForestClassifier
from sklearn.model_selection import cross_val_score, RepeatedStratifiedKFold

%precision 3


SEED = 42


df_train = pd.read_csv("./df/EDA_09_MLP_train.csv")
df_test = pd.read_csv("./df/EDA_09_MLP_test.csv")
Target = 'Perished'
data_cleaner = [df_train, df_test]
df_train.columns


for dataset in data_cleaner:
    dataset = dataset.drop(columns=["Name", "Sex", "Age", "IsAlone", "Ticket", 
                           "FareBin_Code", "Cabin",  "Title", "FareBin", "AgeBin"], inplace=True)



df_train.head()

USE_Columns = df_train.columns
X = df_train.iloc[:, 2:].values
y = df_train.iloc[:, 1].values
X_test = df_test.iloc[:, 1:]


accuracy = {}
speed = {}


model = RandomForestClassifier()
name = "RndomForestClassifierr"
start = time()
cv = RepeatedStratifiedKFold(n_splits=5, n_repeats=5, random_state=SEED)
score = cross_val_score(model, X, y, scoring="accuracy", cv=cv, n_jobs=-1)
speed[name] = np.round(time()-start, 3)
accuracy[name] = np.mean(score).round(3)
print("Model:{}\n".format(name))
print(
    f"Mean Accuracy: {accuracy[name]}\nStd: {np.std(score):.3f}\nRun time: {speed[name]}s"
)


rfc = RandomForestClassifier(max_depth=10,min_samples_leaf=1,n_estimators=100,n_jobs=-1,random_state=SEED)
rfc.fit(X,y)
print('Train Score: {}'.format(round(rfc.score(X, y), 3)))
print(' Test Score: {}'.format(round(rfc.score(, y), 3)))
