==> models/04_Model.ipynb <==
import sys
import os
# データ加工・処理・分析ライブラリ
from sklearn.tree import DecisionTreeClassifier
from sklearn.linear_model import LogisticRegression
from sklearn.linear_model import Lasso
from sklearn.linear_model import Ridge
from sklearn.linear_model import LinearRegression
from sklearn.ensemble import RandomForestClassifier
from sklearn.preprocessing import StandardScaler
from sklearn.model_selection import train_test_split
#Common Model Helpers
from sklearn.preprocessing import OneHotEncoder, LabelEncoder
from sklearn import feature_selection
from sklearn import model_selection
from sklearn import metrics
import sklearn
import numpy as np
import numpy.random as random
import scipy as sp
from pandas import Series, DataFrame
import pandas as pd

# 可視化ライブラリ
import matplotlib.pyplot as plt
import matplotlib as mpl
import seaborn as sns
%matplotlib inline

#Common Model Algorithms
from sklearn import svm, tree, linear_model, neighbors, naive_bayes, ensemble, discriminant_analysis, gaussian_process
from xgboost import XGBClassifier


# 小数第3位まで表示
%precision 3

#ignore warnings
import warnings
warnings.filterwarnings('ignore')
print('-'*25)

train_csv = "./df/EDA_04_train.csv"
test_csv = "./df/EDA_04_test.csv"
df_train = pd.read_csv(train_csv)
df_test = pd.read_csv(test_csv)

X = df_train.iloc[:, 2:].values
y = df_train.iloc[:, 1].values

X_test = df_test.iloc[:, 1:].values


df_train.head()

df_test.head()

Target = ['Perished']

#define x variables for original features aka feature selection
data1_x = ['Sex','Pclass', 'Embarked', 'Title','SibSp', 'Parch', 'Age', 'Fare', 'FamilySize', 'IsAlone'] #pretty name/values for charts
data1_x_calc = ['Sex_Code','Pclass', 'Embarked_Code', 'Title_Code','SibSp', 'Parch', 'Age', 'Fare'] #coded for algorithm calculation
data1_xy =  Target + data1_x
print('Original X Y: ', data1_xy, '\n')


#define x variables for original w/bin features to remove continuous variables
data1_x_bin = ['Sex_Code','Pclass', 'Embarked_Code', 'Title_Code', 'FamilySize', 'AgeBin_Code', 'FareBin_Code']
data1_xy_bin = Target + data1_x_bin
print('Bin X Y: ', data1_xy_bin, '\n')


#define x and y variables for dummy features original
data1_dummy = pd.get_dummies(df_train[data1_x])
data1_x_dummy = data1_dummy.columns.tolist()
data1_xy_dummy = Target + data1_x_dummy
print('Dummy X Y: ', data1_xy_dummy, '\n')

ddf = df_train[['Perished','Sex_Code','Pclass', 'Embarked_Code', 'Title_Code', 'FamilySize', 'AgeBin_Code', 'FareBin_Code']]
ddf_test = df_test[['Sex_Code','Pclass', 'Embarked_Code', 'Title_Code', 'FamilySize', 'AgeBin_Code', 'FareBin_Code']]
ddf_test.head()

X = ddf.iloc[:, 1:].values
y = ddf.iloc[:, 0].values

X_test = ddf_test.iloc[:, 0:].values

#ロジステック回帰モデル

# 訓練データとテストデータに分ける
X_train, X_valid, y_train, y_valid = train_test_split(
    X, y, test_size=0.5, random_state=41)


# 標準化処理
sc = StandardScaler()
sc.fit(X_train)
X_train_std = sc.transform(X_train)
X_valid_std = sc.transform(X_valid)
X_test_std = sc.transform(X_test)

# ロジスティック回帰クラスの初期化と学習
lr = LogisticRegression(random_state=42)
lr.fit(X_train_std, y_train)


# # 正解率の表示
print('正解率(Normal)(train):{:.3f}'.format(lr.score(X_train_std, y_train)))
print('正解率(Nomal)(test):{:.3f}'.format(lr.score(X_valid_std, y_valid)))

from sklearn.model_selection import GridSearchCV
param_grid = {'max_depth': [3, 5, 7, 9],
              'min_samples_leaf': [1, 2, 4, 6,8],
              'n_estimators': [50, 100, 200, 300,500]}
rfc_gs = GridSearchCV(RandomForestClassifier(
    n_estimators=200, n_jobs=1, random_state=42), param_grid, cv=5)
rfc_gs.fit(X_train_std, y_train)

print('Best Parameters: {}'.format(rfc_gs.best_params_))
print('CV Score: {}'.format(round(rfc_gs.best_score_, 3)))

rfc_fe = RandomForestClassifier(max_depth=9, min_samples_leaf=1, n_estimators=200, n_jobs=-1, random_state=42)
rfc_fe.fit(X_train_std,y_train)

print('Train Score: {}'.format(round(rfc_fe.score(X_train_std, y_train), 3)))
print(' Test Score: {}'.format(round(rfc_fe.score(X_valid_std, y_valid), 3)))

from sklearn.neural_network import MLPClassifier
mlpc = MLPClassifier(hidden_layer_sizes=(100, 100, 10), random_state=0)
mlpc.fit(X_train_std, y_train)

print('Multilayer Perceptron \n')
print('Train Score: {}'.format(round(mlpc.score(X_train_std, y_train), 3)))
print(' Test Score: {}'.format(round(mlpc.score(X_valid_std, y_valid), 3)))


#predict
lr_predict = lr.predict_proba(X_test_std)
rfc_predict = rfc_fe.predict_proba(X_test)
mlpc_predict = mlpc.predict_proba(X_test_std)
pred_proba = (lr_predict+2*rfc_predict+2*mlpc_predict)/5
predict = pred_proba.argmax(axis=1)
predict


submission = pd.read_csv('gender_submission.csv')
submission['Perished'] = predict
submission.to_csv(os.path.join("sub", '04_submission.csv'), index=False)
submission.head()
