==> models/06_Model.ipynb <==
#load packages
import sys #access to system parameters https://docs.python.org/3/library/sys.html
print("Python version: {}". format(sys.version))

import pandas as pd #collection of functions for data processing and analysis modeled after R dataframes with SQL like features
from pandas import Series, DataFrame
print("pandas version: {}". format(pd.__version__))

import matplotlib #collection of functions for scientific and publication-ready visualization
import matplotlib.pyplot as plt
import matplotlib as mpl
import seaborn as sns
print("matplotlib version: {}". format(matplotlib.__version__))

import numpy as np #foundational package for scientific computing
print("NumPy version: {}". format(np.__version__))

import scipy as sp #collection of functions for scientific computing and advance mathematics
print("SciPy version: {}". format(sp.__version__)) 

import IPython
from IPython import display #pretty printing of dataframes in Jupyter notebook
print("IPython version: {}". format(IPython.__version__)) 

import sklearn #collection of machine learning algorithms
print("scikit-learn version: {}". format(sklearn.__version__))

#misc libraries
import random
import time


#ignore warnings
import warnings
warnings.filterwarnings('ignore')
print('-'*25)
# データインポートライブラリ
from sklearn.ensemble import RandomForestClassifier
from sklearn.model_selection import StratifiedKFold
from sklearn.metrics import roc_curve, auc
import warnings
import requests
import zipfile
import io
import re as re
import string

# データ加工・処理・分析ライブラリ
import scipy as sp
import sweetviz as sv


# 可視化ライブラリ
from pandas_profiling import ProfileReport
%matplotlib inline


#Model
from sklearn import svm, tree, linear_model, neighbors, naive_bayes, ensemble, discriminant_analysis, gaussian_process
from sklearn.preprocessing import LabelEncoder, OneHotEncoder, StandardScaler
from sklearn import feature_selection
from sklearn import model_selection
from sklearn import metrics
from sklearn.ensemble import RandomForestClassifier
from sklearn.model_selection import StratifiedKFold
from xgboost import XGBClassifier
from sklearn.metrics import roc_curve, auc
%precision 3


SEED = 42


def concat_df(train_data, test_data, sort=True, drop=True):
    """
    overview
    >return a  concatenated df of training and test data
    needed
    >pandas as pd
    input
    >train_data{DataFrame}:train_data
    >test_data{DataFrame}:test_data
    >sort{bool}:True or False
    """
    return pd.concat([train_data, test_data], sort=sort).reset_index(drop=drop)


def divide_df(all_data, target, len=890):
    """
    overview
    > Returns divided dfs of training and test set
    needed
    >
    input
    > all_data{DataFrame}:Train+Test
    > target{str} :Target columns
    > len(int):len of train_data
    
    """
    return all_data.loc[:len], all_data.loc[len:].drop([target], axis=1)


df_train = pd.read_csv("./df/EDA_05_train.csv")
df_test = pd.read_csv("./df/EDA_05_test.csv")
Target = 'Survived'
Final_target = 'Perished'
df_test.shape

df_all = concat_df(df_train, df_test)
drop_cols = ['Deck', 'Embarked', 'Family', 'Family_Size', 'Family_Size_Grouped', Target,
             'Name', 'Parch', 'PassengerId', 'Pclass', 'Sex', 'SibSp', 'Ticket', 'Title',
             'Ticket_Survival_Rate', 'Family_Survival_Rate', 'Ticket_Survival_Rate_NA', 'Family_Survival_Rate_NA']

df_all.drop(columns=drop_cols, inplace=True)

df_all.head()

X_train = StandardScaler().fit_transform(df_train.drop(columns=drop_cols))
y_train = df_train[Target].values
X_test = StandardScaler().fit_transform(df_test.drop(columns=drop_cols))

print('X_train shape: {}'.format(X_train.shape))
print('y_train shape: {}'.format(y_train.shape))
print('X_test shape: {}'.format(X_test.shape))

#Machine Learning Algorithm (MLA) Selection and Initialization
MLA = [
    #Ensemble Methods
    ensemble.AdaBoostClassifier(),
    ensemble.BaggingClassifier(),
    ensemble.ExtraTreesClassifier(),
    ensemble.GradientBoostingClassifier(),
    ensemble.RandomForestClassifier(),

    #Gaussian Processes
    gaussian_process.GaussianProcessClassifier(),
    
    #GLM
    linear_model.LogisticRegressionCV(),
    linear_model.PassiveAggressiveClassifier(),
    linear_model.RidgeClassifierCV(),
    linear_model.SGDClassifier(),
    linear_model.Perceptron(),
    
    #Navies Bayes
    naive_bayes.BernoulliNB(),
    naive_bayes.GaussianNB(),
    
    #Nearest Neighbor
    neighbors.KNeighborsClassifier(),
    
    #SVM
    svm.SVC(probability=True),
    svm.NuSVC(probability=True),
    svm.LinearSVC(),
    
    #Trees    
    tree.DecisionTreeClassifier(),
    tree.ExtraTreeClassifier(),
    
    #Discriminant Analysis
    discriminant_analysis.LinearDiscriminantAnalysis(),
    discriminant_analysis.QuadraticDiscriminantAnalysis(),

    
    #xgboost: http://xgboost.readthedocs.io/en/latest/model.html
    XGBClassifier()    
    ]



#split dataset in cross-validation with this splitter class: http://scikit-learn.org/stable/modules/generated/sklearn.model_selection.ShuffleSplit.html#sklearn.model_selection.ShuffleSplit
#note: this is an alternative to train_test_split
cv_split = model_selection.ShuffleSplit(n_splits = 10, test_size = .3, train_size = .6, random_state = 0 ) # run model 10x with 60/30 split intentionally leaving out 10%

#create table to compare MLA metrics
MLA_columns = ['MLA Name', 'MLA Parameters','MLA Train Accuracy Mean', 'MLA Test Accuracy Mean', 'MLA Test Accuracy 3*STD' ,'MLA Time']
MLA_compare = pd.DataFrame(columns = MLA_columns)

#create table to compare MLA predictions
MLA_predict = y_train

#index through MLA and save performance to table
row_index = 0
for alg in MLA:

    #set name and parameters
    MLA_name = alg.__class__.__name__
    MLA_compare.loc[row_index, 'MLA Name'] = MLA_name
    MLA_compare.loc[row_index, 'MLA Parameters'] = str(alg.get_params())
    
    #score model with cross validation: http://scikit-learn.org/stable/modules/generated/sklearn.model_selection.cross_validate.html#sklearn.model_selection.cross_validate
    cv_results = model_selection.cross_validate(alg, X_train, y_train, cv  = cv_split,return_train_score=True)

    MLA_compare.loc[row_index, 'MLA Time'] = cv_results['fit_time'].mean()
    MLA_compare.loc[row_index, 'MLA Train Accuracy Mean'] = cv_results['train_score'].mean()
    MLA_compare.loc[row_index, 'MLA Test Accuracy Mean'] = cv_results['test_score'].mean()   
    #if this is a non-bias random sample, then +/-3 standard deviations (std) from the mean, should statistically capture 99.7% of the subsets
    MLA_compare.loc[row_index, 'MLA Test Accuracy 3*STD'] = cv_results['test_score'].std()*3   #let's know the worst that can happen!
    

    #save MLA predictions - see section 6 for usage
    alg.fit(X_train, y_train)
    
    row_index+=1

    
#print and sort table: https://pandas.pydata.org/pandas-docs/stable/generated/pandas.DataFrame.sort_values.html
MLA_compare.sort_values(by = ['MLA Test Accuracy Mean'], ascending = False, inplace = True)
MLA_compare
#MLA_predict


print(MLA_compare.loc[21,:])
print(MLA_compare.loc[21,"MLA Parameters"])

Pick_up_Model = [
    discriminant_analysis.LinearDiscriminantAnalysis(),
    XGBClassifier(),
    ensemble.GradientBoostingClassifier(),    
    ]
i =0
cv_split = model_selection.ShuffleSplit(
        n_splits=10, test_size=.3, train_size=.6, random_state=0)

Model1 = Pick_up_Model[0]
cv_results = model_selection.cross_validate(
    Model1, X_train, y_train, cv=cv_split, return_train_score=True)
Model1.fit(X_train,y_train) 
Model1_pred = Model1.predict_proba(X_test)

Model2 = Pick_up_Model[1]
cv_results = model_selection.cross_validate(
    Model2, X_train, y_train, cv=cv_split, return_train_score=True)
Model2.fit(X_train, y_train)
Model2_pred = Model2.predict_proba(X_test)

Model3 = Pick_up_Model[2]
cv_results = model_selection.cross_validate(
    Model3, X_train, y_train, cv=cv_split, return_train_score=True)
Model3.fit(X_train, y_train)
Model3_pred = Model3.predict_proba(X_test)


pred_proba = (Model3_pred)/1
type(pred_proba)
predict = pred_proba.argmax(axis=1)
predict

submission_ex = pd.read_csv("gender_submission.csv")
submission_df = pd.DataFrame(columns=['PassengerId', Target])
submission_df['PassengerId'] = df_test['PassengerId']
submission_df.reset_index(drop=True, inplace=True)
submission_df[Target] = predict
submission_df.to_csv('sub/06_submissions_kaggle.csv', header=True, index=False)
print("Submission Format:{} ".format(submission_ex.shape))
print("My Sumbisison Format:{}".format(submission_df.shape))
submission_df



submission_ex = pd.read_csv("gender_submission.csv")
submission_df = pd.DataFrame(columns=['PassengerId', Target])
submission_df['PassengerId'] = df_test['PassengerId']
submission_df[Target] = predict
submission_df[Final_target] = 0
submission_df.loc[submission_df[Target] == 0,Final_target] = 1
submission_df.loc[submission_df[Target] == 1, Final_target] = 0
submission_df.drop(Target, axis=1, inplace=True)

submission_df.reset_index(drop=True, inplace=True)
submission_df.to_csv('sub/06_submissions.csv', header=True, index=False)
print("Submission Format:{} ".format(submission_ex.shape))
print("My Sumbisison Format:{}".format(submission_df.shape))
submission_df
