==> models/08_Model.ipynb <==
#load packages
import sklearn  # collection of machine learning algorithms
from IPython import display  # pretty printing of dataframes in Jupyter notebook
# collection of functions for scientific computing and advance mathematics
import scipy as sp
import numpy as np  # foundational package for scientific computing
# collection of functions for scientific and publication-ready visualization
import matplotlib
import pandas as pd  # collection of functions for data processing and analysis modeled after R dataframes with SQL like features
from xgboost import XGBClassifier
from sklearn import metrics
from sklearn import model_selection
from sklearn import feature_selection
from sklearn.preprocessing import LabelEncoder, OneHotEncoder, StandardScaler
from sklearn import svm, tree, linear_model, neighbors, naive_bayes, ensemble, discriminant_analysis, gaussian_process
import torch
from torch.utils.data import DataLoader,TensorDataset
from torch.autograd import  Variable
import torch.nn.functional as F
import torch.nn as nn
import torch.optim as optim
import re as re
import io
import zipfile
from sklearn.model_selection import train_test_split
import requests
from sklearn.metrics import roc_curve, auc
from sklearn.model_selection import StratifiedKFold,GridSearchCV
from sklearn.ensemble import RandomForestClassifier
import warnings
import time
import random
import IPython
import seaborn as sns
import matplotlib as mpl
import matplotlib.pyplot as plt
from pandas import Series, DataFrame
import sys  # access to system parameters https://docs.python.org/3/library/sys.html
print("Python version: {}". format(sys.version))

print("pandas version: {}". format(pd.__version__))

print("matplotlib version: {}". format(matplotlib.__version__))

print("NumPy version: {}". format(np.__version__))

print("SciPy version: {}". format(sp.__version__))

print("IPython version: {}". format(IPython.__version__))

print("scikit-learn version: {}". format(sklearn.__version__))

#misc libraries


#ignore warnings
warnings.filterwarnings('ignore')
print('-'*25)
# データインポートライブラリ

# データ加工・処理・分析ライブラリ


# 可視化ライブラリ
%matplotlib inline


#Model
%precision 3


SEED = 42


def concat_df(train_data, test_data, sort=True, drop=True):
    """
    overview
    >return a  concatenated df of training and test data
    needed
    >pandas as pd
    input
    >train_data{DataFrame}:train_data
    >test_data{DataFrame}:test_data
    >sort{bool}:True or False
    """
    return pd.concat([train_data, test_data], sort=sort).reset_index(drop=drop)


def divide_df(all_data, target, len=890):
    """
    overview
    > Returns divided dfs of training and test set
    needed
    >
    input
    > all_data{DataFrame}:Train+Test
    > target{str} :Target columns
    > len(int):len of train_data
    
    """
    return all_data.loc[:len], all_data.loc[len:].drop([target], axis=1)


df_train = pd.read_csv("./df/EDA_05_train.csv")
df_test = pd.read_csv("./df/EDA_05_test.csv")
Target = 'Survived'
Final_target = 'Perished'
df_test.shape

df_all = concat_df(df_train, df_test)
drop_cols = ['Deck', 'Embarked', 'Family', 'Family_Size', 'Family_Size_Grouped', Target,
             'Name', 'Parch', 'PassengerId', 'Pclass', 'Sex', 'SibSp', 'Ticket', 'Title',
             'Ticket_Survival_Rate', 'Family_Survival_Rate', 'Ticket_Survival_Rate_NA', 'Family_Survival_Rate_NA']

df_all.drop(columns=drop_cols, inplace=True)

X = StandardScaler().fit_transform(df_train.drop(columns=drop_cols))
y = df_train[Target].values
X_test = StandardScaler().fit_transform(df_test.drop(columns=drop_cols))
print('X_train shape: {}'.format(X.shape))
print('y_train shape: {}'.format(y.shape))
print('X_test shape: {}'.format(X_test.shape))

X_train, X_valid, y_train, y_valid = train_test_split(
    X, y, test_size=0.5, random_state=SEED)


# XGBClassifier

#Model1 XGBClassifier

param_XGB = {
    'n_estimators': [10, 50, 100, 200], 'early_stopping_rounds':[5,10,20,]
}
XGB = XGBClassifier(booster='gbtree',
                   random_state=SEED, n_estimators=10, early_stopping_rounds=5)


XGB_GSCV= GridSearchCV(XGB,param_grid=param_XGB,scoring='accuracy',cv=5,return_train_score=True)
XGB_GSCV.fit(X_train,y_train,eval_set=[(X_valid,y_valid)])
print('best score: {:0.3f}'.format(XGB_GSCV.score(X_valid, y_valid)))
print('best params: {}'.format(XGB_GSCV.best_params_))
print('best val score:  {:0.3f}'.format(XGB_GSCV.best_score_))


XGB_pred = XGB_GSCV.predict(X_test)
XGB_pred

# GradientBoosting

#Model1 GradientBoosting
from sklearn.ensemble import GradientBoostingClassifier

param_GBC = {
    'n_estimators': [10, 50, 100, 200],
    'early_stopping_rounds': [5, 10, 20, 50],
    'max_depth':[5,10,30,50,100],
    'max_features' :[3,5,10,15]

}
GBC = GradientBoostingClassifier(learning_rate=1.0e-2,n_estimators=5,max_depth=5,max_features=5)

GBC_GSCV = GridSearchCV(XGB, param_grid=param_GBC,
                        scoring='accuracy', cv=5, return_train_score=True)
GBC_GSCV.fit(X_train, y_train, eval_set=[(X_valid, y_valid)])
print('best score: {:0.3f}'.format(XGB_GSCV.score(X_valid, y_valid)))
print('best params: {}'.format(XGB_GSCV.best_params_))
print('best val score:  {:0.3f}'.format(XGB_GSCV.best_score_))


submission_ex = pd.read_csv("gender_submission.csv")
submission_df = pd.DataFrame(columns=['PassengerId', Target])
submission_df['PassengerId'] = df_test['PassengerId']
submission_df[Target] = XGB_pred
submission_df[Final_target] = 0
submission_df.loc[submission_df[Target] == 0, Final_target] = 1
submission_df.loc[submission_df[Target] == 1, Final_target] = 0
submission_df.drop(Target, axis=1, inplace=True)

submission_df.reset_index(drop=True, inplace=True)
submission_df.to_csv('sub/08_submissions.csv', header=True, index=False)
print("Submission Format:{} ".format(submission_ex.shape))
print("My Sumbisison Format:{}".format(submission_df.shape))
submission_df


submission_ex = pd.read_csv("gender_submission.csv")
submission_df = pd.DataFrame(columns=['PassengerId', Target])
submission_df['PassengerId'] = df_test['PassengerId']
submission_df.reset_index(drop=True, inplace=True)
submission_df[Target] = XGB_pred
submission_df.to_csv('sub/08_submissions_kaggle.csv', header=True, index=False)
print("Submission Format:{} ".format(submission_ex.shape))
print("My Sumbisison Format:{}".format(submission_df.shape))
submission_df


X_train_torch = torch.from_numpy(X_train).float()
y_train_torch = torch.from_numpy(y_train).float()
X_valid_torch = torch.from_numpy(X_valid).float()
y_valid_torch = torch.from_numpy(y_valid).float()



type(X_train_torch)

class Model(torch.nn.Module):
    def __init__(self):
        super(Model, self).__init__()
        self.input_size = len(df_train.columns)
        self.fc1 = torch.nn.Linear(self.input_size,128)
        self.fc2 = torch.nn.Linear(128,2)


    def forward(self, x):
        x1 = F.relu(self.fc1(x))
        x2 = self.fc2(x1)
        return F.sigmoid(x2)

        outputs = torch.sigmoid(self.linear(x))
        return outputs


TorchModel = Model()
criterion = nn.CrossEntropyLoss()
optimizer = optim.Adam(TorchModel.parameters(), lr=1.0e-3)
device = torch.device("cuda:0" if torch.cuda.is_available() else "cpu")
print(device)


train_torch = TensorDataset(X_train_torch, y_train_torch)
train_loader = DataLoader(train_torch, batch_size=15, shuffle=True)
param_torch = {
    'epochs':2,
}
for epoch in range(param_torch['epochs']):
    print("Epoch:",epoch)
    total_loss = 0
    for x_train,y_train  in train_loader:
        x_train ,y_train = Variable(x_train),Variable(y_train)
        optimizer.zero_grad()
        output = TorchModel(x_train)
        loss = criterion(output,y_train)
        loss.backward()
        optimizer.step()
        total_loss += loss.item()
