==> models/09_Model_GBDT.ipynb <==
#load packages
import sklearn  # collection of machine learning algorithms
from IPython import display  # pretty printing of dataframes in Jupyter notebook
# collection of functions for scientific computing and advance mathematics
import scipy as sp
import numpy as np  # foundational package for scientific computing
# collection of functions for scientific and publication-ready visualization
import matplotlib
import pandas as pd  # collection of functions for data processing and analysis modeled after R dataframes with SQL like features
from sklearn import metrics
from sklearn import model_selection
from sklearn import feature_selection
from sklearn.preprocessing import LabelEncoder, OneHotEncoder, StandardScaler
from sklearn import svm, tree, linear_model, neighbors, naive_bayes, ensemble, discriminant_analysis, gaussian_process
import torch
from torch.utils.data import DataLoader, TensorDataset
from torch.autograd import Variable
import torch.nn.functional as F
import torch.nn as nn
import torch.optim as optim
import re as re
import io
import zipfile
from sklearn.model_selection import train_test_split
import requests
from sklearn.metrics import roc_curve, auc,log_loss
from sklearn.model_selection import StratifiedKFold, GridSearchCV
from sklearn.ensemble import RandomForestClassifier
import warnings
import time
import random
import IPython
import seaborn as sns
from hyperopt import hp
import xgboost as xgb
from pandas import Series, DataFrame
import sys  # access to system parameters https://docs.python.org/3/library/sys.html
print("Python version: {}". format(sys.version))

print("pandas version: {}". format(pd.__version__))

print("matplotlib version: {}". format(matplotlib.__version__))

print("NumPy version: {}". format(np.__version__))

print("SciPy version: {}". format(sp.__version__))

print("IPython version: {}". format(IPython.__version__))

print("scikit-learn version: {}". format(sklearn.__version__))

#misc libraries


#ignore warnings
warnings.filterwarnings('ignore')
print('-'*25)
# データインポートライブラリ

# データ加工・処理・分析ライブラリ


# 可視化ライブラリ
%matplotlib inline


#Model
%precision 3


SEED = 42


# GBDT

GBDT_train = pd.read_csv("./df/EDA_09_GBDT_train.csv")
GBDT_test = pd.read_csv("./df/EDA_09_GBDT_test.csv")
Target = 'Perished'
GBDT_test.shape
data_cleaner = [GBDT_train, GBDT_test]

GBDT_test.isnull().sum()

for dataset in data_cleaner:
    dataset = dataset.drop(columns=["Name", "Sex", "Age", "IsAlone", "Ticket", "SibSp", "Parch",
                           "FareBin_Code", "Cabin", "Embarked", "Title", "FareBin", "AgeBin"], inplace=True)


GBDT_train.columns

USE_Columns = GBDT_train.columns
X = GBDT_train.iloc[:,2:].values
y = GBDT_train.iloc[:,1].values
X_test = GBDT_test.iloc[:,1:].values
# print('X_train shape: {}'.format(GBDT_X.shape))
# print('y_train shape: {}'.format(GBDT_y.shape))
# print('X_test shape: {}'.format(GBDT_X_test.shape))


X_train, X_valid, y_train, y_valid = train_test_split(
    X, y, test_size=0.5, random_state=SEED)


#Model1 XGBoost
'''
Note
- Tuning はhyperoptで
- 決定木の数は十分大きくしてearly_stoppingにより抑制
- learning rate(eta)は学習時に0.1，提出時により小さくする．
- Turing時のKFoldはCrossValidateionにより1に，実際のモデル作成時にはSEEDにより分割
- uniform は一様分布から抽出する．
- quniformは一様分布から一定の間隔ごとに抽出する．
- logunigoromは対数が一様分布に従う分布から抽出する．


'''
params = {
    'booster':'gbtree',
    'objective':'binary:logistic',
    'eta':0.1,
    'gamma':0.0,
    'alpha':0.0,
    'lambda':1.0,
    'min_child_weight':1,
    'max_depth':5,
    'subsample':0.8,
    'colsample_bytree':0.8,
    'random_state':SEED
}
param_space = {
    'min_child_weight':hp.loguniform('min_child_weight',np.log(0.1),np.log(10)),
    'max_depth':hp.quniform('max_depth',3,9,1),
    'subsample':hp.quniform('subsample',0.6,0.95,0.05),
    'colsample_bytree':hp.quniform('colsample_bytree',0.6,0.95,0.05),
    'gamma':hp.loguniform('gamma',np.log(1e-8),np.log(1.0)),
    #余裕があれば
    # 'alpha':hp.loguniform('alpha',np.log(1e-8),np.log(1.0)),
    # 'lambda': hp.loguniform('alpha', np.log(1e-8), np.log(1.0)),

}
# XGBoodt 特徴量選択
dtrain = xgb.DMatrix(X_train, label=y_train)
dvalid = xgb.DMatrix(X_valid,label= y_valid)
dtest = xgb.DMatrix(X_test)
watchlist = [(dtrain,'train'),(dvalid,'eval')]
num_round = 50



def score(params):
    

model = xgb.train(params, dtrain, num_round, evals=watchlist)
va_pred = model.predict(dvalid)
score = log_loss(y_valid,va_pred)
print(score)


pred = model.predict(dvalid)
pred

submission_ex = pd.read_csv("gender_submission.csv")
submission_df = pd.DataFrame(columns=['PassengerId', Target])
submission_df['PassengerId'] = GBDT_test['PassengerId']
submission_df[Target] = XGB_pred
submission_df.reset_index(drop=True, inplace=True)
submission_df.to_csv('sub/09_GBDT_submissions.csv', header=True, index=False)
print("Submission Format:{} ".format(submission_ex.shape))
print("My Sumbisison Format:{}".format(submission_df.shape))
submission_df.head(20)


submission_df[Target].value_counts()
