==> models/10_RandomForest.ipynb <==
#Sytem
import sys,os
import numpy as np
import pandas as pd
import warnings
import time
import random
import IPython
warnings.filterwarnings('ignore')

#Visualidation
import sweetviz as sv
import matplotlib.pyplot as plt
import matplotlib
%matplotlib inline

#EDA
import sweetviz as sv
import sklearn
from sklearn.preprocessing import OneHotEncoder, LabelEncoder



#model
from sklearn.ensemble import RandomForestClassifier
from sklearn.model_selection import train_test_split,RandomizedSearchCV,GridSearchCV
from skopt import BayesSearchCV




print('-'*40)

print("Python version: {}". format(sys.version))

print("pandas version: {}". format(pd.__version__))

print("matplotlib version: {}". format(matplotlib.__version__))

print("NumPy version: {}". format(np.__version__))


print("scikit-learn version: {}". format(sklearn.__version__))
print('-'*40)


# Info

Target = "Perished"
PATH_LOG = "./log.csv"
PATH_df = "./df"
ID = 10
SEED = 42

# Import Dataset

df_train = pd.read_csv("../data/input/train.csv")
df_test = pd.read_csv("../data/input/test.csv")
#log = pd.read_csv(PATH_LOG)
data = pd.concat([df_train,df_test],sort=False)
print('-'*25)
print("Shape of Train Data\n[row :{},column :{}]".format(
    df_train.shape[0], df_train.shape[1]))
print('-'*25)
print("Shape of Test Data\n[row :{},column :{}]".format(
    df_test.shape[0], df_test.shape[1]))
print('-'*25)

log.tail(2)

# EDA

## Report

# repoort = sv.analyze(df_train,target_feat="Perished")
# repoort.show_html("SWEETVIZ_REPORT.html")

## Feature Engineering

label = LabelEncoder()
ohe = OneHotEncoder()

### Title

data['Title'] = data['Name'].str.split(
        ", ", expand=True)[1].str.split(".", expand=True)[0]

title_names = (data.Title.value_counts() < 10)
data["Title"] = data["Title"].apply(lambda x: "other" if title_names.loc[x] == True else x)

### Age

#for title in data.Title.unique():
#    mean = data.loc[(data.Title.values == title),"Age"].mean()
#    data.loc[(data.Title.values == title) & (data.Age.isnull()),"Age"] = mean


### Fare

faremean = data.Fare.mean()
data["Fare"].fillna(faremean,inplace=True)
data['FareBin'] = pd.qcut(data['Fare'], 10)
data['Fare_Code'] = label.fit_transform(data['FareBin'])

### Encoding

N_Age_Code = 3
data.replace({'Sex':{'male':0,'female':1}},inplace=True)
data['Age_Code'] = label.fit_transform(pd.qcut(data['Age'], N_Age_Code))
data['Embarked_Code'] = label.fit_transform(data['Embarked'])
data["Title_Code"] = label.fit_transform(data["Title"])

### Embarked

data["Embarked"].fillna('S',inplace=True)
embarked_ohe = pd.get_dummies(data.Embarked)
data = pd.concat([data, embarked_ohe], axis=1)

data["Family"] = data["SibSp"] + data["Parch"]

drop_columns = ["Name","Ticket","Cabin","Embarked","Title",'Embarked_Code','FareBin',"Fare_Code"]
dataset = data.drop(columns=drop_columns)
train = dataset[:len(df_train)]
test = dataset[len(df_train):]

data.Age.isnull().sum()

test.head()

train.to_csv(os.path.join(PATH_df, "EDA_"+str(ID)+"_train.csv"), index_label=False)
test.to_csv(os.path.join(PATH_df,"EDA_"+str(ID)+ "_test.csv"), index_label=False)

# Model

X = train.iloc[:, 2:].values
y = train.iloc[:, 1].values
X_test = test.iloc[:, 2:].values

test.isnull().sum()

X_train, X_valid, y_train, y_valid = train_test_split(X, y, test_size=0.2, random_state=42)

base_model = RandomForestClassifier(n_estimators = 10, random_state = SEED)
base_model.fit(X_train, y_train)
print('Train Score: {}'.format(round(base_model.score(X_train, y_train), 4)))
print(' Valid Score: {}'.format(round(base_model.score(X_valid, y_valid), 4)))

Modelname = "RFC"
rfc = RandomForestClassifier()
CV = 3
n_estimators = [int(x) for x in np.linspace(start = 200, stop = 2000, num = 10)]
max_features = ['auto', 'sqrt']
max_depth = [int(x) for x in np.linspace(10, 110, num = 11)]
max_depth.append(None)
min_samples_split = [2, 5, 10]
min_samples_leaf = [1, 2, 4]
bootstrap = [True, False]
param_grid = {
    'n_estimators':n_estimators,
    'max_features':max_features,
    'max_depth':max_depth,
    'min_samples_split':min_samples_split,
    'min_samples_leaf':min_samples_leaf,
    'bootstrap':bootstrap
}
print(param_grid)

## Param Search

### Grid Search

# Search = "GS"
# rfc_cv = GridSearchCV(estimator=rfc,param_grid=param_grid,scoring='accuracy',cv=CV,verbose=2,n_jobs=-1)
# rfc_cv.fit(X_train,y_train)

### Random Search

Search = "RS"
rfc_cv = RandomizedSearchCV(estimator=rfc,param_distributions=param_grid,n_iter=100,cv=CV,verbose=1,random_state=SEED,n_jobs=-1)
rfc_cv.fit(X_train,y_train)

### Bays Search

Search = "BS"
rfc_cv = BayesSearchCV(estimator=rfc,search_spaces=param_grid,n_iter=100,cv=CV,verbose =2)
rfc_cv.fit(X,y)

print(rfc_cv.best_params_)

best_model = rfc_cv.best_estimator_
print(best_model)
print('Train Score: {}'.format(round(best_model.score(X_train, y_train), 4)))
print(' Valid Score: {}'.format(round(best_model.score(X_valid, y_valid), 4)))

bestparam = str(rfc_cv.best_params_)
STrain = best_model.score(X_train, y_train)
SValid  = best_model.score(X_valid, y_valid)

## Predict

pred = rfc_cv.predict(X_test)

## Submit

PATH_sub =  ("./sub/"+str(ID)+"_submission_"+Modelname+".csv")
print(PATH_sub)
submission_ex = pd.read_csv("gender_submission.csv")
submission_df = pd.DataFrame(columns=['PassengerId', Target])
submission_df['PassengerId'] = df_test['PassengerId']
submission_df[Target] = pred
submission_df.reset_index(drop=True, inplace=True)
submission_df.to_csv(PATH_sub, header=True, index=False)
print("Submission Format:{} ".format(submission_ex.shape))
print("My Sumbisison Format:{}".format(submission_df.shape))
submission_df.head(5)

## Log

Row  =  len(log)
log.loc[Row] = "x"
log.loc[Row,"Modelname"] = Modelname
log.loc[Row,"Search"] = Search
log.loc[Row,"CV"] = CV
log.loc[Row, "Bestparams"] = bestparam 
log.loc[Row,"STrain"] = STrain
log.loc[Row,"SValid"] = SValid
log.loc[Row,"Age_Code_N"] = N_Age_Code
use_columns = test.columns.drop(["PassengerId","Perished"])
for col in use_columns:
    log.loc[Row,col] = "o"
log.fillna("x",inplace=True)
# if PublicScore is knowed
#log.loc[Row,"SPublic"] = 0.791

log.to_csv(PATH_LOG,index_label=False)
log.tail()
